UnliftIO.MessageBox.CatchAll

Description

Utilities for exception safe message boxes.

This provides a wrapper around UnliftIO.MessageBox.Class instances to catch SomeException in all methods like deliver and receive.

Documentation

newtype CatchAllFactory cfg

A wrapper around values that are instances of IsMessageBoxFactory. The factory wraps the result of the delegated newMessageBox invocation into a CatchAllBox.

Constructors

CatchAllFactory cfg

newtype CatchAllBox box a

A wrapper around values that are instances of IsMessageBox.

The Input type will be wrapped using CatchAllInput.

Constructors

CatchAllBox (box a)

newtype CatchAllInput i a

A wrapper around values that are instances of IsInput.

Constructors

CatchAllInput (i a)