UnliftIO.MessageBox.Class

Description

This module contains a type class that describes exchangable operations on messages boxes.

Documentation

class (IsMessageBox (MessageBox cfg), IsInput (Input (MessageBox cfg))) => IsMessageBoxFactory cfg where

Create IsMessageBox instances from a parameter. Types that determine MessageBox values.

For a limited message box this might be the limit of the message queue.

Associated Types

type MessageBox cfg :: Type -> Type

Methods

getConfiguredMessageLimit :: cfg -> Maybe Int

Return a message limit.

NOTE: This method was added for unit tests. Although the method is totally valid, it might not be super useful in production code. Also note that the naming follows the rule: Reserve short names for entities that are used often.

newMessageBox :: MonadUnliftIO m => cfg -> m (MessageBox cfg a)

Create a new msgBox. This is required to receive a message. NOTE: Only one process may receive on an msgBox.

class IsInput (Input msgBox) => IsMessageBox msgBox where

A type class for msgBox types. A common interface for receiving messages.

Minimal complete definition

receive, tryReceive, newInput

Associated Types

type Input msgBox :: Type -> Type

Type of the corresponding input

Methods

receive :: MonadUnliftIO m => msgBox a -> m (Maybe a)

Receive a message. Take whatever time it takes. Return Just the value or Nothing when an error occurred.

NOTE: Nothing may sporadically be returned, especially when there is a lot of load, so please make sure to build your application in such a way, that it anticipates failure.

tryReceive :: MonadUnliftIO m => msgBox a -> m (Future a)

Return a Future that can be used to wait for the arrival of the next message. NOTE: Each future value represents the next slot in the queue so one future corresponds to exactly that message (should it arrive) and if that future value is dropped, that message will be lost!

receiveAfter

Arguments

:: MonadUnliftIO m

=> msgBox a
Message box

-> Int
Time in micro seconds to wait until the action is invoked.

-> m (Maybe a)

Wait for an incoming message or return Nothing.

The default implementation uses tryReceive to get a Future on which awaitFuture inside a timeout is called.

Instances might override this with more performant implementations especially non-blocking Unagi channel based implementation.

NOTE: Nothing may sporadically be returned, especially when there is a lot of load, so please make sure to build your application in such a way, that it anticipates failure.

newInput :: MonadUnliftIO m => msgBox a -> m (Input msgBox a)

Create a new input that enqueus messages, which are received by the msgBox

class IsInput input where

A type class for input types. A common interface for delivering messages.

Methods

deliver :: MonadUnliftIO m => input a -> a -> m Bool

Send a message. Take whatever time it takes. Depending on the implementation, this might be a non-blocking operation. Return if the operation was successful.

NOTE: False may sporadically be returned, especially when there is a lot of load, so please make sure to build your application in such a way, that it anticipates failure.

handleMessage :: (MonadUnliftIO m, IsMessageBox msgBox) => msgBox message -> (message -> m b) -> m (Maybe b)

Receive a message and apply a function to it.