UnliftIO.MessageBox.Limited

Description

Thread safe queues for uni directional message passing between threads.

This message box has an upper limit, that means that sometimes delivery either fails or is blocked until the receiving thread has consumed more messages.

Use this module if the producer(s) outperform the consumer, but you want the extra safety that the queue blocks the Input after a certain message limit is reached.

If you are sure that the producers fire at a slower rate then the rate at which the consumer consumes messages, use this module.

Documentation

data MessageLimit

Message Limit

The message limit must be a reasonable small positive integer that is also a power of two. This stems from the fact that Unagi is used under the hood.

The limit is a lower bound.

Constructors

MessageLimit_1

MessageLimit_2

MessageLimit_4

MessageLimit_8

MessageLimit_16

MessageLimit_32

MessageLimit_64

MessageLimit_128

MessageLimit_256

MessageLimit_512

MessageLimit_1024

MessageLimit_2048

MessageLimit_4096

messageLimitToInt :: MessageLimit -> Int

Convert a MessageLimit to the Int representation.

newtype BlockingBoxLimit

Contains the (vague) limit of messages that a BlockingBox can buffer, i.e. that deliver can put into a BlockingInput of a BlockingBox.

Constructors

BlockingBoxLimit MessageLimit

data BlockingBox a

A message queue out of which messages can by received.

This is the counter part of Input. Can be used for reading messages.

Messages can be received by receive or tryReceive.

data BlockingInput a

A message queue into which messages can be enqued by, e.g. tryToDeliver. Messages can be received from an BlockingBox.

The Input is the counter part of a BlockingBox.

newtype NonBlockingBoxLimit

A BlockingBoxLimit wrapper for non-blocking IsMessageBoxArg instances.

Constructors

NonBlockingBoxLimit MessageLimit

data NonBlockingBox a

A BlockingBox wrapper for non-blocking IsMessageBox instances.

The difference to the BlockingBox instance is that deliver immediately returns if the message box limit is surpassed.

newtype NonBlockingInput a

A wrapper around BlockingInput with a non-blocking IsInput instance.

deliver will enqueue the message or return False immediately, if the message box already contains more messages than it's limit allows.

Constructors

NonBlockingInput (BlockingInput a)

data WaitingBoxLimit

A IsMessageBoxArg instance wrapping the BlockingBox with independently configurable timeouts for receive and deliver.

Constructors

WaitingBoxLimit !(Maybe Int) !Int !MessageLimit

data WaitingBox a

A BlockingBox an a WaitingBoxLimit for the IsMessageBox instance.

Constructors

WaitingBox WaitingBoxLimit (BlockingBox a)

data WaitingInput a

An input for a BlockingBox that will block for not much more than the given timeout when the message box is full.

Constructors

WaitingInput !Int !(BlockingInput a)