UnliftIO.MessageBox.Unlimited

Description

Thread safe queues for message passing between many concurrent processes.

This message box is UNLIMITED.

Good single producer/single consumer performance

If you are sure that the producer(s) send messages at a lower rate than the rate at which the consumer consumes messages, use this module.

Otherwise use the more conservative UnliftIO.MessageBox.Limited module.

Documentation

data BlockingUnlimited

The (empty) configuration for creating UnlimitedBoxes using the IsMessageBoxFactory methods.

Constructors

BlockingUnlimited

data UnlimitedBox a

A message queue out of which messages can by received.

This is the counter part of Input. Can be used for reading messages.

Messages can be received by receive or tryReceive.

data UnlimitedBoxInput a

A message queue into which messages can be enqued by, e.g. deliver. Messages can be received from an UnlimitedBox.

The UnlimitedBoxInput is the counter part of a UnlimitedBox.