UnliftIO.MessageBox.Util.Fresh

Description

Threadsafe, shared, atomic counters

This is based on Data.Atomics.Counter.

Documentation

fresh :: forall a env m. (MonadReader env m, MonadIO m, HasCounterVar a env, Coercible a Int) => m a

A threadsafe atomic a

Atomically increment and get the value of the Counter for type a that must be present in the env.

incrementAndGet :: forall a m. (MonadIO m, Coercible a Int) => CounterVar a -> m a

Atomically increment and get the value of the Counter for type a that must be present in the env.

newCounterVar :: forall a m. MonadIO m => m (CounterVar a)

Create a new CounterVar starting at 0.

class HasCounterVar a env | env -> a where

A type class for MonadReader based applications.

Methods

getCounterVar :: env -> CounterVar a

data CounterVar a

An AtomicCounter.