UnliftIO.MessageBox.Util.Future

Documentation

newtype Future a

A wrapper around an IO action that returns value in the future.

Constructors

Future (IO (Maybe a))

tryNow :: MonadUnliftIO m => Future a -> m (Maybe a)

Return Just the value or Nothing, when the value is not available yet.

Once the value is available, that value will be returned everytime this function is invoked.

awaitFuture :: MonadUnliftIO m => Future b -> m b

Poll a Future until the value is present.